chore(codeownersProcessor): increase test coverage pt.1
